growth_prediction/requirements.txt
# requirements.txt
# =============================
# PyTorch 관련 라이브러리
torch>=2.0.0
torchvision>=0.15.0

# 데이터 처리 및 분석
pandas>=2.0.0
numpy>=1.24.0

# 이미지 처리
Pillow>=9.5.0

# 머신러닝 관련
scikit-learn>=1.3.0

# 시간 및 날짜 관련
pytz>=2023.3

# 시각화 라이브러리
matplotlib>=3.7.0
